refactor: migrate ChatViewModel to Hilt
